Revert "kill unused include"
This reverts commit 1d329832ed144d9eb802a5b405220dd86cf9a14a.

+++ position.h
@@ -22,9 +22,7 @@
 #ifndef POSITION_H_INCLUDED_
 #define POSITION_H_INCLUDED_
 
-#include <QString>   // for QString
-#include <QVector>   // for QVector
-#include <QtGlobal>  // for qint64
+#include <QVector>         // for QVector
 
 #include "defs.h"    // for route_head (ptr only), ARG_NOMINMAX, ARGTYPE_FLOAT
 #include "filter.h"  // for Filter
@@ -45,7 +43,7 @@ private:
   route_head* cur_rte = nullptr;
 
   double pos_dist{};
-  qint64 max_diff_time{};
+  double max_diff_time{};
   char* distopt = nullptr;
   char* timeopt = nullptr;
   char* purge_duplicates = nullptr;
